Only one sound Pool now
The drum pad is now constant through out on beat or one track
We won't be using different drum pads with different sounds assigned to
their buttons for the four patterns. It's the same drum pad and same set
of 16 sounds through out. The sounds can be changed using the edit drum
machine button.